Update README.md
Adding some warning notices to the README so that people know that there's weird stuff for a few cases. This is based on calculations I was doing and getting weird results.

This whole thing isn't user-friendly, it's not even Phil-friendly, but it works for the moment until I can spend some time making it less of an abomination.